Refactor: Splitting handleItemInfo to utility file
